Refactor: Rename useSlider hook to useSelectedSnapDisplay
