+++ src/hooks/useSelectedSnapDisplay.tsx
